feat: 제품 조회 관련 기능 구현

+++ src/main/java/com/homesweet/homesweetback/domain/product/product/controller/response/ProductResponse.java
@@ -54,7 +54,7 @@ public record ProductResponse(
     public record SkuResponse(
             Long id,
             Integer priceAdjustment,
-            Integer stockQuantity,
+            Long stockQuantity,
             Integer finalPrice,
             boolean inStock,
             List<Long> optionValueIndexes
